refactor: move files within its own directories
