+++ src/FormComponents/Radio/Radio.scss
@@ -1,7 +1,7 @@
 @import '../../Variables.scss';
 
 fieldset {
-  border: 2px solid $color-grey-light;
+  border: 1px solid $color-grey-light;
 }
 
 label {
