+++ src/FormComponents/Select/Select.js
@@ -1,5 +1,6 @@
 import { useField } from 'formik';
-
+import clsx from 'clsx';
+import './Select.scss';
 export default function Select({ label, htmlFor, children, ...props }) {
   const [field, meta] = useField(props);
 
@@ -8,7 +9,16 @@ export default function Select({ label, htmlFor, children, ...props }) {
       <label htmlFor={htmlFor} className='label select'>
         {label}
       </label>
-      <select {...field} {...props} className='input'>
+      <select
+        {...field}
+        {...props}
+        className={clsx({
+          input: true,
+          'input-error': meta.error,
+          success: !meta.error,
+          touched: meta.touched,
+        })}
+      >
         {children}
       </select>
       {meta.touched && meta.error && <div className='error'>{meta.error}</div>}
