+++ src/SignupForm/SignupForm.scss
@@ -1,20 +1,47 @@
 @import '../Variables.scss';
 
 .form {
-  padding: 1rem;
+  padding: 4rem;
+  padding-bottom: 6rem;
+  max-width: 800px;
   display: flex;
   flex-direction: column;
-  gap: 1rem;
-  max-width: 500px;
+  gap: 2.5rem;
+  color: $color-grey;
 
-  .input {
-    display: block;
-    // background-color: aqua;
-    width: 100%;
+  select {
+    font-family: $font-main;
   }
 
-  .label {
-    // background-color: red;
-    // margin-bottom: 0.5rem;
+  option {
+    font-family: $font-main;
+  }
+
+  h2 {
+    margin-bottom: 2rem;
+  }
+}
+
+.form__column {
+  display: flex;
+  flex-direction: column;
+  gap: 2.5rem;
+}
+
+.form__title {
+  font-size: $font-size-max;
+}
+
+.error {
+  margin-top: 0.5rem;
+  font-size: 0.8rem;
+  color: $color-error;
+}
+
+/* Media Queries */
+@media (min-width: $breakpoint-tablet) {
+  .nombres {
+    display: flex;
+    gap: $gap;
   }
 }
